+++ .storybook/main.js
@@ -19,19 +19,23 @@ const config = {
     options: {},
   },
 
-  staticDirs: ['../assets'],
+  staticDirs: ["../assets"],
 
   async viteFinal(config) {
-    config.optimizeDeps.exclude = [...(config.optimizeDeps?.exclude ?? []), 'lit', 'lit-html']
+    config.optimizeDeps.exclude = [
+      ...(config.optimizeDeps?.exclude ?? []),
+      "lit",
+      "lit-html",
+    ];
     config.resolve.alias = {
       ...config.resolve.alias,
-      '@sebastienbarbier/design-system': "../src",
+      "@sebastienbarbier/design-system": "../src",
     };
     return config;
   },
 
   docs: {
     autodocs: true,
-  }
+  },
 };
 export default config;