+++ .storybook/manager.js
@@ -1,5 +1,5 @@
-import { addons } from '@storybook/manager-api';
-import yourTheme from './theme';
+import { addons } from "@storybook/manager-api";
+import yourTheme from "./theme";
 
 addons.setConfig({
   theme: yourTheme,