+++ .storybook/theme.js
@@ -1,9 +1,9 @@
-import { create } from '@storybook/theming/create';
+import { create } from "@storybook/theming/create";
 
 export default create({
-  base: 'light',
-  brandTitle: 'sebastienbarbier - Design System',
-  brandUrl: '/',
-  brandImage: '/images/sebastienbarbier_logo.svg',
-  brandTarget: '_self',
+  base: "light",
+  brandTitle: "sebastienbarbier - Design System",
+  brandUrl: "/",
+  brandImage: "/images/sebastienbarbier_logo.svg",
+  brandTarget: "_self",
 });