+++ stories/utils/fromJsToCssVar.js
@@ -1,16 +1,16 @@
-export default function fromJsToCssVar(jsVar, prefix = '--') {
+export default function fromJsToCssVar(jsVar, prefix = "--") {
+  let result = [];
 
-    let result = [];
+  const keys = Object.keys(jsVar);
 
-    const keys = Object.keys(jsVar);
+  keys.forEach((key) => {
+    if (typeof jsVar[key] === "object") {
+      result = [...result, ...fromJsToCssVar(jsVar[key], prefix + key + "-")];
+    } else if (prefix !== "--") {
+      // Skip first layer
+      result.push(`${prefix}${key}`);
+    }
+  });
 
-    keys.forEach(key => {
-        if (typeof jsVar[key] === 'object') {
-            result = [...result, ...fromJsToCssVar(jsVar[key], prefix + key + '-')]
-        } else if (prefix !== '--') { // Skip first layer
-            result.push(`${prefix}${key}`);
-        }
-    });
-
-    return result;
+  return result;
 }