+++ CHANGELOG.md
@@ -1,10 +1,10 @@
-
 # Change log
+
 All notable changes to this project will be documented in this file.
- 
+
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
- 
+
 <!---
 ## [Unreleased] - yyyy-mm-dd
 
@@ -22,20 +22,28 @@ See for sample https://raw.githubusercontent.com/favoloso/conventional-changelog
 -->
 
 ## [0.3.0] - 2024-MM-DD
+
 ### 🛠 Improvements
+
 - Add prettier for code formatting (#11)
 - Add LICENCE to release (#12)
 - Update dependencies (#10)
 
 ## [0.2.0] - 2024-06-24
+
 ### ✨ Feature
+
 - Remove dist from SCSS @use path (#4)
 - Add Badge component (#6)
+
 ### 🐛 Bug Fixes
+
 - List of CSS variables not working on Chrome and Safari (#5)
 
 ## [0.1.0] - 2024-06-17
+
 ### ✨ Feature
+
 - Publish @sebastienbarbier/design-system to npmjs (#3)
-- Deploy storybook to github.io (#2) 
+- Deploy storybook to github.io (#2)
 - Add Storybook documentation (#1)